+++ src/views/issuance/components/zapV2/submit/ZapConfirmButton.tsx
@@ -1,76 +1,22 @@
 import { LoadingButton } from 'components/button'
-import { useEffect } from 'react'
-import {
-  Address,
-  usePrepareSendTransaction,
-  useSendTransaction,
-  useWaitForTransaction,
-} from 'wagmi'
-import { ZapErrorType, useZap } from '../context/ZapContext'
 import TransactionsIcon from 'components/icons/TransactionsIcon'
 import { Box, Spinner, Text } from 'theme-ui'
+import { useZap } from '../context/ZapContext'
+import { useZapTx } from '../context/ZapTxContext'
 import ZapGasCost from '../overview/ZapGasCost'
 
-type ZapConfirmButtonProps = {
-  hasAllowance: boolean
-  loadingApproval: boolean
-  approvalSuccess: boolean
-  setError: (error?: ZapErrorType) => void
-}
-
-const ZapConfirmButton = ({
-  hasAllowance,
-  loadingApproval,
-  approvalSuccess,
-  setError,
-}: ZapConfirmButtonProps) => {
-  const { zapResult, chainId, operation } = useZap()
-
-  const { config } = usePrepareSendTransaction(
-    zapResult
-      ? {
-          data: zapResult.tx.data as Address,
-          gas: BigInt(zapResult.gas),
-          to: zapResult.tx.to as Address,
-          value: BigInt(zapResult.tx.value),
-        }
-      : undefined
-  )
-
+const ZapConfirmButton = () => {
+  const { operation, zapResult } = useZap()
   const {
-    data,
-    isLoading: loadingTx,
+    hasAllowance,
+    loadingApproval,
+    approvalSuccess,
+    loadingTx,
+    validatingTx,
     sendTransaction,
-    error: sendError,
-  } = useSendTransaction(config)
-
-  const {
-    data: receipt,
-    isLoading: validatingTx,
-    error: validatingTxError,
-  } = useWaitForTransaction({
-    hash: data?.hash,
-    chainId,
-  })
-
-  useEffect(() => {
-    if (approvalSuccess && sendTransaction) {
-      sendTransaction()
-    }
-  }, [approvalSuccess, sendTransaction])
-
-  useEffect(() => {
-    if (sendError || validatingTxError) {
-      setError({
-        title: 'Transaction rejected',
-        message: 'Please try again',
-        color: 'danger',
-        secondaryColor: 'rgba(255, 0, 0, 0.20)',
-      })
-    } else {
-      setError(undefined)
-    }
-  }, [sendError, validatingTxError, setError])
+    receipt,
+    error,
+  } = useZapTx()
 
   if (
     (loadingApproval ||
@@ -78,8 +24,7 @@ const ZapConfirmButton = ({
       loadingTx ||
       validatingTx ||
       receipt) &&
-    !sendError &&
-    !validatingTxError
+    !error
   ) {
     return (
       <Box variant="layout.verticalAlign">
